Clova-FinAgent/stock_analyzer/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dateutil==2.8.2
pydantic==2.5.0
langchain-google-genai==1.0.10
langchain-core==0.1.52
PyYAML==6.0.1
pandas==2.1.4